+++ src/GenericWebsocketClientWorker.h
@@ -20,9 +20,12 @@
 #ifndef WSCPP_GENERICWEBSOCKETCLIENTWORKER_H
 #define WSCPP_GENERICWEBSOCKETCLIENTWORKER_H
 
+#include <cstdint>
 #include <memory>
+#include <string>
 #include <tuple> // for std::ignore
 
+#include <boost/variant.hpp>
 #include <nan.h>
 #include <uv.h>
 
@@ -36,6 +39,20 @@
 namespace wscpp
 {
 
+namespace events
+{
+struct Open {
+};
+struct Close {
+  int code;
+  std::string reason;
+};
+struct Error {
+  int code;
+  std::string reason;
+};
+} // namespace events
+
 /**
  * This class is modeled after NaN::AsyncWorker and
  * NaN::AsyncProgressWorkerBase.
@@ -47,9 +64,41 @@ template <typename Config>
 class GenericWebsocketClientWorker : public IWebsocketClientWorker
 {
 private:
-  enum class EventCallbackType { None, Open, Error, Close };
+  using Event = boost::variant<events::Open, events::Close, events::Error>;
   using Endpoint = websocketpp::client<Config>;
   using MessagePtr = typename Config::message_type::ptr;
+
+  class EventVisitor : public boost::static_visitor<>
+  {
+  public:
+    EventVisitor(GenericWebsocketClientWorker* worker, v8::Local<v8::Object> context)
+        : worker(worker), context(context)
+    {
+    }
+
+    void operator()(const events::Open& e) const
+    {
+      worker->parameters->onOpenCallback->Call(context, 0, nullptr);
+    }
+
+    void operator()(const events::Close& e) const
+    {
+      v8::Local<v8::Value> argv[] = {Nan::New<v8::Int32>(e.code),
+                                     Nan::New<v8::String>(e.reason).ToLocalChecked()};
+      worker->parameters->onCloseCallback->Call(context, 2, argv);
+    }
+
+    void operator()(const events::Error& e) const
+    {
+      v8::Local<v8::Value> argv[] = {Nan::New<v8::Int32>(e.code),
+                                     Nan::New<v8::String>(e.reason).ToLocalChecked()};
+      worker->parameters->onErrorCallback->Call(context, 2, argv);
+    }
+
+  private:
+    GenericWebsocketClientWorker* worker;
+    v8::Local<v8::Object> context;
+  };
 
 protected:
   using ConnectionHandle = websocketpp::connection_hdl;
@@ -107,11 +156,11 @@ public:
     endpoint.send(connectionHandle, data, size, websocketpp::frame::opcode::binary);
   }
 
-  void close() override
+  void close(std::uint16_t code, const std::string& reason) override
   {
     endpoint.stop_perpetual();
     websocketpp::lib::error_code ec;
-    endpoint.close(connectionHandle, websocketpp::close::status::going_away, "", ec);
+    endpoint.close(connectionHandle, code, reason, ec);
     if (ec) {
       const std::string errorMessage = "could not close connection: " + ec.message();
       Nan::ThrowError(errorMessage.c_str());
@@ -167,7 +216,7 @@ public:
   void handReceivedMessagesToNode()
   {
     Nan::HandleScope scope;
-    auto context = Nan::New(parameters->thisContext);
+    v8::Local<v8::Object> context = Nan::New(parameters->thisContext);
     MessagePtr msg;
     while (receivedMessageQueue.try_dequeue(msg)) {
       const websocketpp::frame::opcode::value opcode = msg->get_opcode();
@@ -175,16 +224,15 @@ public:
         v8::Local<v8::Value> argv[] = {Nan::New<v8::String>(msg->get_payload()).ToLocalChecked()};
         parameters->onMessageCallback->Call(context, 1, argv);
       } else if (opcode == websocketpp::frame::opcode::binary) {
-        // TODO potentially one copy could be avoided here
-        // this requires that websocketpp provides the possibility to move out
-        // the payload from websocketpp::message_buffer
-        // OR the shared_ptr is kept alive as long as the node::Buffer lives
-        // char* rawData = const_cast<char*>(msg->get_payload().data());
-        // v8::Local<v8::Value> argv[] = {
-        //        Nan::NewBuffer(rawData,
-        //        msg->get_payload().size()).ToLocalChecked()};
-        v8::Local<v8::Value> argv[] = {
-            Nan::CopyBuffer(msg->get_payload().data(), msg->get_payload().size()).ToLocalChecked()};
+        // move payload into a heap allocated string
+        // node Buffer wraps around that string's data
+        // this string will be deleted when the node Buffer goes out of scope
+        auto* payload = new std::string(std::move(msg->get_raw_payload()));
+        auto buffer = Nan::NewBuffer(const_cast<char*>(payload->data()), payload->size(),
+                                     bufferFreeCallback<std::string>, payload)
+                          .ToLocalChecked();
+
+        v8::Local<v8::Value> argv[] = {std::move(buffer)};
         parameters->onMessageCallback->Call(context, 1, argv);
       } else {
         // TODO error handling?
@@ -195,22 +243,11 @@ public:
   void handReceivedEventsToNode()
   {
     Nan::HandleScope scope;
-    auto context = Nan::New(parameters->thisContext);
-    EventCallbackType event;
+    v8::Local<v8::Object> context = Nan::New(parameters->thisContext);
+    Event event;
+    EventVisitor visitor(this, context);
     while (eventQueue.try_dequeue(event)) {
-      switch (event) {
-      case EventCallbackType::Open:
-        parameters->onOpenCallback->Call(context, 0, nullptr);
-        break;
-      case EventCallbackType::Close:
-        parameters->onCloseCallback->Call(context, 0, nullptr);
-        break;
-      case EventCallbackType::Error:
-        parameters->onErrorCallback->Call(context, 0, nullptr);
-        break;
-      default:
-        break;
-      }
+      boost::apply_visitor(visitor, event);
     }
   }
 
@@ -255,20 +292,21 @@ private:
   void connectionOpened(ConnectionHandle connectionHandle)
   {
     this->connectionHandle = std::move(connectionHandle);
-    signalEventOccurred(EventCallbackType::Open);
+    signalEventOccurred(events::Open());
   }
 
   void connectionClosed(ConnectionHandle connectionHandle)
   {
-    std::ignore = connectionHandle;
-    signalEventOccurred(EventCallbackType::Close);
+    auto connection = endpoint.get_con_from_hdl(connectionHandle);
+    signalEventOccurred(
+        events::Close{connection->get_remote_close_code(), connection->get_remote_close_reason()});
   }
 
   void connectionFailed(ConnectionHandle connectionHandle)
   {
-    std::ignore = connectionHandle;
-    // TODO add error description to event
-    signalEventOccurred(EventCallbackType::Error);
+    auto connection = endpoint.get_con_from_hdl(connectionHandle);
+    auto errorCode = connection->get_ec();
+    signalEventOccurred(events::Error{errorCode.value(), errorCode.message()});
   }
 
   /**
@@ -279,10 +317,16 @@ private:
   /**
    * @brief notify node to call `eventOccuredCallback`
    */
-  void signalEventOccurred(EventCallbackType event)
+  void signalEventOccurred(Event event)
   {
-    eventQueue.enqueue(event);
+    eventQueue.enqueue(std::move(event));
     uv_async_send(eventAsyncHandle);
+  }
+
+  template <typename T>
+  static void bufferFreeCallback(char* data, void* hint)
+  {
+    delete reinterpret_cast<T*>(hint);
   }
 
   uv_work_t work;
@@ -292,7 +336,7 @@ private:
   ConnectionHandle connectionHandle;
 
   moodycamel::ConcurrentQueue<MessagePtr> receivedMessageQueue;
-  moodycamel::ConcurrentQueue<EventCallbackType> eventQueue;
+  moodycamel::ConcurrentQueue<Event> eventQueue;
 };
 
 } // namespace wscpp