+++ wiki/ReleaseNotes.md
@@ -1,2 +1,11 @@
+# wscpp 0.2.0
+
+* Added support for `code` and `reason` parameters for `close()`.
+* Added support for specifying `binary` option when sending messages.
+* `onclose` callback is invoked with event containing `code` and `reason`.
+* `onerror` callback is invoked with event containing `code` and `reason`.
+* Made require mechanism of native module part more robust.
+
 # wscpp 0.1.0
+
 This is the initial release of wscpp.